refactoring source file locations and build script
